+++ utils/filesystem/filepath.go
@@ -29,3 +29,8 @@ func FileTreeDepth(fs FS, root, filePath string) (depth int64, err error) {
 	depth = int64(len(strings.Split(diff, "/")) - 1)
 	return
 }
+
+// EndsWithPathSeparator states whether a path is ending with a path separator of not
+func EndsWithPathSeparator(fs FS, filePath string) bool {
+	return strings.HasSuffix(filePath, "/") || strings.HasSuffix(filePath, string(fs.PathSeparator()))
+}